requirements.txt
bs4
click
jinja2
markdown-it-py[plugins]
python-frontmatter
